requirements.txt
selenium == 4.24.0
pytest == 8.3.2
mypy == 1.11.2
flake8 == 7.1.1
black == 24.8.0
isort == 5.13.2
PyHamcrest == 2.1.0
allure-pytest ==2.13.5